% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWKMT2.R
\name{AWKMT2cpp}
\alias{AWKMT2cpp}
\title{Adaptively Weighted Kaplan-Meier Tests}
\usage{
AWKMT2cpp(indata, tau, c_first = 0, c_last = 4, c_by = 0.1,
  method = "permutation", nmethod = 10000, seed = 1, test = "1_side")
}
\arguments{
\item{indata}{A data matrix (data frame). The 1st column is time-to-event variable, the 2nd column is event indicator (1=event, 0=censor), and the 3rd column is the treatment indicator (1=treat, 0=control). No missing values are allowed in this data matrix.}

\item{tau}{A numeric value to specify the time interval of interest. The end of study time will be a general choice.}

\item{c_first}{A first number in range to specify the search area of "c" for the versatile tests by Uno et al. (2015). Default is \code{0}.}

\item{c_last}{A last number in range to specify the search area of "c" for the versatile tests by Uno et al. (2015). Default is \code{4}.}

\item{c_by}{A number to specify the search area of "c" for the versatile tests by Uno et al. (2015). Default is \code{0.1}.}

\item{method}{A name of the resampling method. It supports \code{"permutation"} (default) and \code{"perturbation"}.}

\item{nmethod}{A number of iterations for the resampling. Recommended to specify at least \code{10000} (default) or larger.}

\item{seed}{An integer value, used for the random number generation in the resampling procedures. Default is \code{1}.}

\item{test}{Specify \code{"1_side"} for the one-sided test where the alternative hypothesis is that treatment group is superior to control group with respect to survival.
Specify \code{"2_side"} for the two-sided test where the alternative hypothesis is that treatment group is not equal to control group with respect to survival.
Default is \code{"1_side"}.}
}
\value{
A list with components:

\item{test_method}{One-sided or two-sided test}

\item{resampling_method}{The resampling method.}

\item{crude_pvalue_V1}{Crude p-value of the test based on v1 in Uno et al. (2015).}

\item{crude_pvalue_V2}{Crude p-value of the test based on v2 in Uno et al. (2015).}

\item{bona_fide_pvalue_V1}{Bona-fide p-value of the test based on v1 in Uno et al. (2015).}

\item{bona_fide_pvalue_V2}{Bona-fide p-value of the test based on v2 in Uno et al. (2015).}
}
\description{
Performs the two-sample tests based on adaptively weighted differences between two Kaplan-Meier curves proposed by Uno, Tian, Claggett and Wei (2015).
}
\references{
Uno H, Tian L, Claggett B, Wei LJ. A versatile test for equality of two survival functions based on weighted differences of Kaplan-Meier curves.
 Statistics in Medicine 2015, 34, 3680-3695.
}
